Change the setpoints to be the stop point.
Previously if the setpoint was 70* with a 2* tolerance, heat would turn on at 68 and bring the temp to 70. Instead, we will now turn on at 70 and bring the tem to 72. This makes more sense when we talk about both Heating and Cooling. If the heating target is 70* and the cooling target is 75*, this thermostat will now ensure the temperature stays within 70 and 75.